Port Nanochat Lookup (#1978)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Port nanochat lookup from Delta-v.
Pr: https://github.com/DeltaV-Station/Delta-v/pull/2794

---

# TODO

<!--
A list of everything you have to do before this PR is "complete"
You probably won't have to complete everything before merging but it's
good to leave future references
-->

- [x] Task
- [x] Completed Task

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![Captura de tela 2025-03-16
183052](https://github.com/user-attachments/assets/2b844906-4117-4560-9045-69b9d5493cf3)
![Captura de tela 2025-03-16
183117](https://github.com/user-attachments/assets/b2546750-c808-468e-8fbf-de1d75402c06)
![Captura de tela 2025-03-16
183124](https://github.com/user-attachments/assets/01c5eb53-b1ba-4ca7-8a2e-f1813147b441)
![Captura de tela 2025-03-16
183147](https://github.com/user-attachments/assets/304ac291-b5b0-474f-84c3-8321386c02e2)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Toby222 and Will-Oliver-Br
add: Add Number lookup to NanoChat

---------

Co-authored-by: Tobias Berger <toby@tobot.dev>