Lavaland Shuttle Fix (#1948)
# Description

The bug in question is the one where if you return from Lavaland the
shuttle would not dock to the station but instead send you approximately
around the station.

I replaced the original function and lookup completely where when the
station and lavaland are initialized it will add them to destinations
and then use a list to warp to the gridUIDs. This allows for more
consistent warping and less time for looking up what the largest grid
is. This fix also fixes the issue of smaller maps like Northway where a
wreck could be much larger than the station and would not warp to it.


Here is the new format for the mining shuttle console

![image](https://github.com/user-attachments/assets/31fa9e54-5033-46ca-90fc-ae916e9638fa)
---

# TODO

- [x] Rework the shuttle, adding destinations to make it more modular. 

---

# Changelog

:cl:
- fix: NT has finally patched the mining shuttle and now docks to the
station on returning from lavaland.

---------

Co-authored-by: Nathaniel Adams <60526456+Nathaniel-Adams@users.noreply.github.com>